refactor: encapsulate key-value pair utilities within KeyValueEditor component
