+++ src/editor/utils.js
@@ -373,129 +373,3 @@ export const sanitizePreparedSchemaString = ( schemaStr ) => {
 		return schemaStr;
 	}
 };
-
-/**
- * Utility functions for managing key-value pairs.
- */
-
-/**
- * Add a new key-value pair to the list if no empty pairs exist.
- *
- * @param {Array} list - Array of [key, value] pairs
- * @param {Function} updateFunction - Function to update the list state
- * @return {void}
- */
-export const addKeyValuePair = ( list, updateFunction ) => {
-	// Check if there's already an empty pair
-	if ( list.some( ( [ key, value ] ) => key === '' && value === '' ) ) {
-		return;
-	}
-	// Add new empty pair
-	updateFunction( [ ...list, [ '', '' ] ] );
-};
-
-/**
- * Update a specific key-value pair in the list.
- *
- * @param {Array} list - Array of [key, value] pairs
- * @param {Function} updateFunction - Function to update the list state
- * @param {number} index - Index of the pair to update
- * @param {string} field - Field to update ('key' or 'value')
- * @param {string} fieldValue - New value for the field
- * @return {void}
- */
-export const updateKeyValuePair = (
-	list,
-	updateFunction,
-	index,
-	field,
-	fieldValue
-) => {
-	const updatedList = list.map( ( [ key, value ], i ) => {
-		if ( i === index ) {
-			return field === 'key'
-				? [ fieldValue, value ]
-				: [ key, fieldValue ];
-		}
-		return [ key, value ];
-	} );
-	updateFunction( updatedList );
-};
-
-/**
- * Remove a key-value pair from the list by index.
- *
- * @param {Array} list - Array of [key, value] pairs
- * @param {Function} updateFunction - Function to update the list state
- * @param {number} index - Index of the pair to remove
- * @return {void}
- */
-export const removeKeyValuePair = ( list, updateFunction, index ) => {
-	updateFunction( list.filter( ( _, i ) => i !== index ) );
-};
-
-/**
- * Filter out empty key-value pairs (where either key or value is empty/whitespace).
- *
- * @param {Array} list - Array of [key, value] pairs
- * @param {Function} updateFunction - Function to update the list state
- * @return {void}
- */
-export const filterEmptyKeyValuePairs = ( list, updateFunction ) => {
-	const filtered = list.filter( ( [ key, value ] ) => {
-		const keyStr = String( key || '' ).trim();
-		const valueStr = String( value || '' ).trim();
-		return keyStr !== '' && valueStr !== '';
-	} );
-	updateFunction( filtered );
-};
-
-/**
- * Check if the add button should be disabled.
- * Disabled when the last entry has an empty key or value.
- *
- * @param {Array} list - Array of [key, value] pairs
- * @return {boolean} Whether the add button should be disabled
- */
-export const isAddButtonDisabled = ( list ) => {
-	return (
-		list.length > 0 &&
-		( list[ list.length - 1 ][ 0 ] === '' ||
-			list[ list.length - 1 ][ 1 ] === '' )
-	);
-};
-
-/**
- * Convert key-value pairs array to object and filter out empty pairs.
- *
- * @param {Array} list - Array of [key, value] pairs
- * @return {Object} Object with filtered key-value pairs
- */
-export const keyValuePairsToObject = ( list ) => {
-	if ( ! Array.isArray( list ) ) {
-		return {};
-	}
-
-	const filtered = list.filter( ( item ) => {
-		// Ensure item is an array with at least 2 elements
-		if ( ! Array.isArray( item ) || item.length < 2 ) return false;
-
-		const [ key, value ] = item;
-
-		// Handle null/undefined values
-		if ( key == null || value == null ) return false;
-
-		try {
-			// Convert to string safely and trim
-			const keyStr = String( key ).trim();
-			const valueStr = String( value ).trim();
-			return keyStr !== '' && valueStr !== '';
-		} catch ( error ) {
-			// If String conversion fails for some reason, skip this item
-			console.warn( 'Error converting key-value pair to string:', error );
-			return false;
-		}
-	} );
-
-	return Object.fromEntries( filtered );
-};
